refactor(rendering): move SvgBoard to svg/board
Made-with: Cursor
