+++ src/components/ARCHITECTURE.md
@@ -58,4 +58,4 @@ These components are typically composed within **Pages** (e.g., `src/pages/MainM
 - **State Management:** The parent page manages the visibility (e.g., `isSettingsOpen`) and data fetching (e.g., list of games from the `Session` model).
 - **Navigation:** Buttons within components (like "Continue" in `GameCard`) trigger callbacks that the parent page uses to initiate navigation via `react-router-dom`.
 
-SVG board tiles (`HexTile`, `SvgBoard`, segment renderers) live under **`src/rendering/svg/tiles/`** for parity with the canvas `TilePreview` in **`src/rendering/canvas/components/TilePreview.tsx`**. Layout uses `SvgHexUtils.hexToPixel`, delegating to **`src/rendering/common/hex/HexUtils.hexToPixel`** with `SVG_HEX_RADIUS`.
+SVG **board** composition (`SvgBoard`) lives under **`src/rendering/svg/board/`**; per-tile primitives (`HexTile`, wedges, segment renderers, `SvgHexUtils`) under **`src/rendering/svg/tiles/`**. Canvas parity uses **`src/rendering/canvas/components/TilePreview.tsx`**. Layout uses `SvgHexUtils.hexToPixel`, delegating to **`src/rendering/common/hex/HexUtils.hexToPixel`** with `SVG_HEX_RADIUS`.
