+++ src/rendering/svg/board/SvgBoard.stories.tsx
